first draft of breaking up monolithic file into constituents
