+++ gp/__init__.py
@@ -5,174 +5,11 @@ from pathlib import Path
 import numpy as np
 import uproot
 import hist
-from sklearn.gaussian_process import GaussianProcessRegressor, kernels
 from scipy.stats import chisquare
-import mplhep
-mplhep.style.use('ROOT')
 
-import matplotlib.pyplot as plt
-
-
-def sim_imd():
-    """simulate the IMD by sampling from a moyal distribution
-
-    The values of the two moyal distribution parameters were taken
-    from a fit of a Moyal PDF to the 6.5% 2016 IMD. The returned
-    histogram has the same binning as this IMD and roughly the same
-    number of entries (~5M, some lost to overflow).
-
-    Returns
-    -------
-    hist.Hist
-        simulated IMD
-    """
-    import scipy
-    return (
-        hist.Hist.new
-        .Reg(6000, 0.0, 0.3, label = 'Mass / GeV')
-        .Double()
-        .fill(
-            scipy.stats.moyal.rvs(
-                loc = 0.065,
-                scale = 0.013,
-                size = 5_000_000
-            )
-        )
-    )
-
-
-def load_imd(fp : str|Path, imd_name : str = 'invM_h'):
-    """Load an IMD from the input file and make sure the x axis is labeled appropriately
-
-    Parameters
-    ----------
-    filepath: str|Path
-        path to ROOT file to load
-    imd_name: str, optional, default invM_h
-        key name of histogram in ROOT file
-
-    Returns
-    -------
-    hist.Hist
-        loaded IMD from the input file
-    """
-    with uproot.open(fp) as f:
-        h = f[imd_name].to_hist()
-        h.axes[0].label = 'Mass / GeV'
-        return h
-
-
-def _deduce_histogram(h: hist.Hist|str):
-    """Deduce and return the histogram that should be used from the input specification
-
-    Meant to be used within the construction of the GP model class below.
-
-    Parameters
-    ----------
-    h: hist.Hist|str
-        If a hist.Hist is given, use that as the histogram.
-        If h is a str, there are two possible values.
-        'sim' returns the result of sim_imd and 'real' returns the result of 'load_imd'
-        with the filepath 'hps2016invMHisto10pc.root'.
-        Any other str produces a ValueError and any other type produces a TypeError.
-
-    Returns
-    -------
-    hist.Hist
-        histogram following input specification
-    """
-
-    if isinstance(h, str):
-        if h == 'sim':
-            return sim_imd()
-        elif h == 'real':
-            # load histogram from ROOT file
-            return load_imd('hps2016invMHisto10pc.root')
-        else:
-            raise ValueError(f'Histogram specification {h} not understood.')
-    elif isinstance(h, hist.Hist):
-        return h
-    else:
-        raise TypeError(f'Histogram specification of type {type(h)} not supported.')
-
-
-def fit(
-    histogram: hist.Hist,
-    kernel,
-    blind_range = None,
-    **kwargs
-) -> GaussianProcessRegressor:
-    """fit the input histogram with a GP using the input kernel
-
-    optionally, blind the fit to a range of the histogram
-
-    Parameters
-    ----------
-    histogram: hist.Hist
-        histogram to fit a GP to
-    kernel:
-        kernel to use in GP
-    blind_range: 2-tuple, optional, default None
-        range of histogram to "blind" the fit to
-        (i.e. do /not/ use this range of values in fit)
-    kwargs: dict[str,Any]
-        all the rest of the keyword arguments are passed to GaussianProcessRegressor
-
-
-    Note
-    ----
-    The running of this fit can take a while. The GP is manipulating N-dimensional
-    matrices where N corresponds to the number of training points. Since there are
-    6000 bins in the input invariant mass histogram, N=6000 which are really large
-    matrices and can take a long time to run.
-    The manipulation of large matrices is a problem _built_ for GPUs and hence
-    we may want to switch to a GPU-possible implementation of GPs like GPyTorch[1].
-    In the mean time, I would highly recommend pickling[2] the resulting fitted GP object
-    so that the fit doesn't need to be re-run if you just want to make different plots
-    with the predictions the GP makes.
-
-        import pickle
-        with open('helpful-name-for-this-gp-fit.pkl','wb') as f:
-            pickle.dump(gp, f)
-
-    and then somewhere else (e.g. in a notebook where you are playing with plots) you can
-
-        import pickle
-        with open('helpful-name-for-this-gp-fit.pkl','rb') as f:
-            gp = pickle.load(f)
-
-
-    [1]: https://docs.gpytorch.ai/en/stable/examples/01_Exact_GPs/Simple_GP_Regression.html
-    [2]: https://docs.python.org/3/library/pickle.html
-    """
-
-    x     = histogram.axes[0].centers
-    value = histogram.values()
-
-    fit_selection = (value > 0.0) # make sure something is in the bins
-    if blind_range is not None:
-        if isinstance(blind_range, (tuple,list)) and len(blind_range)==2:
-            fit_selection = fit_selection&((x < blind_range[0])|(x > blind_range[1]))
-        else:
-            raise ValueError('blind_range is not a length-2 tuple defining the range of coordinates to blind')
-
-    x_train = x[fit_selection]
-    y_train = value[fit_selection]
-    variance = histogram.variances()[fit_selection]
-
-    if 'alpha' in kwargs:
-        raise KeyError('alpha cannot be manually set. It is determined to be the variance of the histogram to fit')
-
-    # update n_restarts_optimizer default to be 9 as is used in example
-    kwargs.setdefault('n_restarts_optimizer', 9)
-    _gp = GaussianProcessRegressor(
-        kernel = kernel,
-        alpha = variance,
-        **kwargs
-    )
-    # fit expects a _column_ for x and a _row_ for y so we need to reshape x
-    _gp.fit(x_train.reshape(-1,1), y_train)
-    return _gp
+from ._fit import fit, kernels
+from ._hist import sim_imd, load, _deduce_histogram, rebin_and_limit, inject_signal
+from ._plot import plt, label
 
 
 class GaussianProcessModel:
@@ -218,9 +55,104 @@ class GaussianProcessModel:
             blind_range = self.blind_range,
             **kwargs
         )
+        self.mean_pred, self.std_pred = self.predict(self.histogram.axes[0].centers)
+        self.chi2_statistic, self.p_value = chisquare(
+            self.histogram.values(),
+            np.sum(self.histogram.values())/np.sum(self.mean_pred) * self.mean_pred
+        )
 
-    def plot(self):
-        """Plot a comparison between the histogram and the (presumed already fit) GP"""
+
+    def predict(self, mass):
+        """Evaluate the model and the input mass, returning its prediction and uncertainty
+        for the event yield
+
+        Parameters
+        ----------
+        mass: np.array
+            array of masses to evaluate in GeV
+
+        Return
+        ------
+        2-tuple of np.array
+            the mean prediction and the uncertainty on that prediction for the event yield
+            at the input masses
+        """
+        return self.model.predict(mass.reshape(-1,1), return_std=True)
+
+    
+    @property
+    def combined_variance(self):
+        return self.histogram.variances()+self.std_pred**2
+
+    @property
+    def positive_prediction(self):
+        return (self.mean_pred > 0)&(self.histogram.values() > 0)
+
+    @property
+    def ratio(self):
+        vals = (
+            self.histogram.values()[self.positive_prediction]
+            /self.mean_pred[self.positive_prediction]
+        )
+        err = vals*np.sqrt(
+            self.std_pred[self.positive_prediction]**2
+            /self.mean_pred[self.positive_prediction]
+        )
+        return vals, err
+
+
+    @property
+    def pull(self):
+        return (
+            (self.histogram.values()-self.mean_pred)[self.positive_prediction]
+            /np.sqrt(self.combined_variance[self.positive_prediction])
+        )
+
+
+    def plot_histogram(self, **kwargs):
+        kwargs.setdefault('flow', None)
+        kwargs.setdefault('label', 'Observed Data')
+        art = self.histogram.plot(
+            **kwargs
+        )
+        kwargs.get('ax', plt.gca()).set_ylabel('Event Yield')
+        return art
+
+
+    def plot_prediction(self, ax = None, **kwargs):
+        kwargs.setdefault(
+            'label',
+            '\n'.join([
+                'GP with 95% Confidence Interval',
+                f'$\chi^2 = ${self.chi2_statistic:.3g}',
+                r'$P_{\chi^2} = $'+f'{self.p_value:.3g}'
+            ])
+        )
+        x = self.histogram.axes[0].centers
+        if ax is None:
+            ax = plt.gca()
+        line_art, = ax.plot(x, self.mean_pred, **kwargs)
+        fill_art = ax.fill_between(
+             x, self.mean_pred - 1.96*self.std_pred, self.mean_pred + 1.96*self.std_pred,
+             alpha = 0.5, color = line_art.get_color()
+        )
+        return line_art, fill_art
+
+
+    def plot_pull_histogram(self, **kwargs):
+        kwargs.setdefault('yerr', False)
+        art = (
+            hist.Hist.new
+            .Reg(41,-10,10,label=r'Pull $(\mathrm{Data}-\mathrm{GP})/\sigma$')
+            .Double()
+            .fill(pull_values)
+        ).plot(**kwargs)
+        kwargs.get('ax', plt.gca()).set_ylabel('Bin Count')
+        return art
+    
+
+    def plot_comparison(self):
+        """Plot a comparison between the histogram and the Fit"""
     
         fig, axes = plt.subplots(
             nrows = 3,
@@ -234,74 +166,36 @@ class GaussianProcessModel:
 
         raw, ratio, pull = axes
 
-        # evaluate GP model prediction
-        x = self.histogram.axes[0].centers
-        mean_pred, std_pred = self.model.predict(x.reshape(-1,1), return_std=True)
-        
-        # CHI2
-        chi2_statistic, p_value = chisquare(self.histogram.values(), np.sum(self.histogram.values())/np.sum(mean_pred) * mean_pred)
-
         # RAW
+        x = self.histogram.axes[0].centers[self.positive_prediction]
         # hist has plotting methods already
         #   add label (for legend) and don't show the flow bins
         #   (default is to draw a little arrow hinting that something exists out there)
-        self.histogram.plot(ax=raw, label='Observed Data\nChi2 = %.3f \nChi2 Probability = %.3f' % (chi2_statistic, p_value), flow=None)
-        art, = raw.plot(
-            x, mean_pred,
-            label= 'GP with 95% Confidence Interval'
-        )
-        raw.fill_between(
-             x, mean_pred - 1.96*std_pred, mean_pred + 1.96*std_pred,
-             alpha = 0.5, color = art.get_color()
-        )
+        self.plot_histogram(ax=raw)
+        self.plot_prediction(ax=raw)
         raw.legend(
-            #title = f'Kernel: {repr(self.model.kernel_)}',
             title = 'Kernel: RBF * DotProduct',
             title_fontsize = 'xx-small'
         )
-        raw.set_ylabel('Event Count')
         # mpl default is to add some horizontal padding which I don't like
         raw.set_xlim(np.min(x), np.max(x))
-        mplhep.label.exp_label('HPS', llabel='Internal', rlabel='2016 6.5%', ax=raw)
-
-        combined_variance = self.histogram.variances()+std_pred**2
-        positive_prediction = (mean_pred > 0)&(self.histogram.values() > 0)
+        label(ax=raw)
 
         # RATIO
-        ratio_values = (
-            self.histogram.values()[positive_prediction]
-            /mean_pred[positive_prediction]
-        )
-        ratio.plot(x[positive_prediction], ratio_values)
-        ratio_err = ratio_values*np.sqrt(
-            std_pred[positive_prediction]**2
-            /mean_pred[positive_prediction]**2
-            +self.histogram.variances()[positive_prediction]
-            /self.histogram.values()[positive_prediction]**2
-        )
-        ratio.fill_between(
-            x[positive_prediction], ratio_values - 1.96*ratio_err, ratio_values + 1.96*ratio_err,
-            alpha = 0.5
-        )
+        ratio_vals, ratio_err = self.ratio
+        ratio.plot(x, ratio_vals)
+        ratio.fill_between(x, ratio_vals - ratio_err, ratio_vals+ratio_err, alpha=0.5)
         ratio.axhline(1, color='gray', ls=':')
-        ratio.set_ylabel(r'Data / GP')
+        ratio.set_ylabel(r'Data / Fit')
 
         # PULL
-        pull_values = (
-            (self.histogram.values()-mean_pred)[positive_prediction]
-            /np.sqrt(combined_variance[positive_prediction])
-        )
-        pull.plot(x[positive_prediction], pull_values)
+        pull.plot(x, self.pull)
         pull.fill_between(
             x, np.full(x.shape, -2), np.full(x.shape, +2),
             color='gray', alpha=0.5
         )
-        pull.set_ylabel(r'$(\mathrm{Data} - \mathrm{GP})/\sigma$')
+        pull.set_ylabel(r'$(\mathrm{Data} - \mathrm{Fit})/\sigma$')
 
-        # CHI2
-        chi2_results = chisquare(self.histogram.values(), np.sum(self.histogram.values())/np.sum(mean_pred) * mean_pred)
-        print(chi2_results)
-        
         # FINAL CLEANUP
 
         for ax in axes[:-1]:
@@ -314,156 +208,3 @@ class GaussianProcessModel:
                     ax.axvline(e, color='tab:red')
 
         return fig, axes
-
-    def get_chisquare(self):
-        x = self.histogram.axes[0].centers
-        mean_pred, std_pred = self.model.predict(x.reshape(-1,1), return_std=True)
-        chi2_statistic, p_value = chisquare(self.histogram.values(), np.sum(self.histogram.values())/np.sum(mean_pred) * mean_pred)
-        return chi2_statistic
-        
-    def get_chisquare_p_value(self):
-        x = self.histogram.axes[0].centers
-        mean_pred, std_pred = self.model.predict(x.reshape(-1,1), return_std=True)
-        chi2_statistic, p_value = chisquare(self.histogram.values(), np.sum(self.histogram.values())/np.sum(mean_pred) * mean_pred)
-        return p_value
-
-    def get_pull_values(self):
-        
-        x = self.histogram.axes[0].centers
-        mean_pred, std_pred = self.model.predict(x.reshape(-1,1), return_std=True)
-        
-        combined_variance = self.histogram.variances()+std_pred**2
-        positive_prediction = (mean_pred > 0)&(self.histogram.values() > 0)
-        
-        pull_values = (
-            (self.histogram.values()-mean_pred)[positive_prediction]
-            /np.sqrt(combined_variance[positive_prediction])
-        )
-        return pull_values[positive_prediction]
-        
-
-    def plot_pull_histogram(self):
-        x = self.histogram.axes[0].centers
-        mean_pred, std_pred = self.model.predict(x.reshape(-1,1), return_std=True)
-        combined_variance = self.histogram.variances()+std_pred**2
-        positive_prediction = (mean_pred > 0)&(self.histogram.values() > 0)
-        pull_values = (
-            (self.histogram.values()-mean_pred)[positive_prediction]
-            /np.sqrt(combined_variance[positive_prediction])
-        )
-        art = (
-            hist.Hist.new
-            .Reg(41,-10,10,label=r'Pull $(\mathrm{Data}-\mathrm{GP})/\sigma$')
-            .Double()
-            .fill(pull_values)
-        ).plot(yerr=False)
-        plt.ylabel('Bin Count')
-        return art
-    
-
-    def plot_raw(self):
-        art = self.histogram.plot(label='Histogram', flow=None)
-        plt.ylabel('Event Count')
-        return art
-
-
-
-@dataclass
-class rebin_and_limit:
-    """rebin an input histogram by the defined factor"""
-    factor: int = None
-    low_limit: float = None
-    high_limit: float = None
-
-
-    def __call__(self, h):
-        return h[slice(
-            hist.loc(self.low_limit) if self.low_limit is not None else None,
-            hist.loc(self.high_limit) if self.high_limit is not None else None,
-            hist.rebin(self.factor) if self.factor > 1 else None
-        )]
-
-
-@dataclass
-class inject_signal:
-    """add a signal bump onto the histogram
-
-    Parameters
-    ----------
-    amplitude: float
-        total number of events to have bump represent
-    width: float
-        width of gaussian signal bump (standard deviation)
-    location: float
-        center of gaussian signal bump (mean)
-    """
-    amplitude: float
-    width: float
-    location: float
-    
-
-    def __call__(self, h):
-        """UNTESTED"""
-        import scipy
-        signal_samples = self.amplitude*scipy.stats.norm.pdf(
-            h.axes[0].centers,
-            loc=self.location,
-            scale=self.width
-        )
-        h[:] = h.values() + signal_samples
-        return h
-
-
-def fit_and_plot(
-    name,
-    blind_range = None,
-    low_lim = 0.033,
-    up_lim = 0.179,
-    rebin = 10,
-    plot_filename = None,
-    display = True
-):
-    gpm = GaussianProcessModel(
-        h = 'real',
-        kernel = 1.0 * kernels.RBF() * kernels.DotProduct(),
-        blind_range = blind_range,
-        modify_histogram = [
-            rebin_and_limit(rebin, low_lim, up_lim),
-            inject_signal(1000.0, 0.100, 0.008)
-        ]
-    )
-    fig, axes = gpm.plot()
-    if plot_filename is not None:
-        fig.savefig(plot_filename, bbox_inches='tight')
-    if display:
-        plt.show()
-    else:
-        fig.clf()
-    return gpm
-
-
-def _cli_main():
-    import argparse
-    parser = argparse.ArgumentParser()
-    parser.add_argument('name', help='name for model pickle and image')
-    parser.add_argument('--blind', nargs=2, type=float, help='bounds of region to blind')
-    parser.add_argument('--rebin', type=int, help='rebin factor', default=10)
-    parser.add_argument('--low-lim', type=float, default=0.033, help='lower limit of fit')
-    parser.add_argument('--up-lim', type=float, default=0.179, help='upper limit of fit')
-    args = parser.parse_args()
-
-    gpm = fit_and_plot(
-        args.name,
-        blind_range = args.blind,
-        rebin = args.rebin,
-        low_lim = args.low_lim,
-        up_lim = args.up_lim,
-        plot_filename = f'{args.name}.png',
-        display = False
-    )
-    with open(f'{args.name}.pkl','wb') as f:
-        pickle.dump(gpm.model, f)
-
-
-if __name__ == '__main__':
-    _cli_main()
